Separate EmissiveOreBakedModel to its own file
